# reflect: optimize Value.IsZero for array types